Changed UI to reflect new style
