--- src/index.css
@@ -1,204 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Nunito+Sans&family=Poppins&display=swap');
-
-body {
-  background: #353b48;
-  margin: 0;
-  padding: 0;
-}
-
-.mobile-view {
-  position: relative;
-  height: 100vh;
-  width: 100vw;
-  float: left;
-  background: #2f3640;
-}
-
-.sidebar {
-  background: #2f3640;
-  height: 100vh;
-  width: 250px;
-  position: relative;
-  float: left;
-}
-
-.controller {
-  background: #2f3640;
-  width: 250px;
-  position:absolute;
-  bottom:0;
-  left:0;
-}
-
-.controller-mobile {
-  background: #2f3640;
-  width: 100%;
-  position:absolute;
-  bottom:0;
-  left:0;
-}
-
-.album-art {
-  width: 250px;
-  height: 250px;
-  object-fit: cover;
-}
-
-.album-art-mobile {
-  width: 64px;
-  height: 64px;
-  object-fit: cover;
-  float: left;
-}
-
-.table {
-  width: 100%;
-}
-
-.song-details-mobile {
-  height: 64px;
-  float: left;
-  width: 50vw;
-}
-
-.song-name {
-  font-size: 18px;
-  font-weight: bold;
-  font-family: 'Nunito Sans', sans-serif;
-  color: white;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -o-text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
-}
-
-.song-name-mobile {
-  font-size: 16px;
-  font-weight: bold;
-  font-family: 'Nunito Sans', sans-serif;
-  color: white;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -o-text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
-}
-
-.artist-name {
-  font-size: 16px;
-  font-family: 'Nunito Sans', sans-serif;
-  color: white;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -o-text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
-}
-
-.artist-name-mobile {
-  font-size: 14px;
-  font-family: 'Nunito Sans', sans-serif;
-  color: white;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -o-text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
-}
-
-.transparent {
-  background: transparent;
-}
-
-.menu-mobile {
-  position: relative;
-  float: left;
-}
-
-.menu-active {
-  color: #FFA600;
-  font-weight: bold;
-  font-family: 'Poppins', sans-serif;
-  border-radius: 8px;
-  cursor: default;
-}
-
-.menu-inactive {
-  color: white;
-  font-family: 'Poppins', sans-serif;
-  border-radius: 8px;
-  cursor: pointer;
-}
-
-.headline-dark {
-  color: #1a1a1a;
-  font-family: 'Poppins', sans-serif;
-  font-weight: bold;
-}
-
-.page-headline {
-  color: white;
-  font-family: 'Poppins', sans-serif;
-  font-weight: bold;
-}
-
-.seekbar {
-  width: 100%;
-  float: left;
-}
-
-.audio-controls {
-  width: 100%
-}
-
-.play-button-mobile {
-  float: right;
-  margin-top: 24px;
-  margin-right: 18px;
-  position: relative;
-}
-
-.navbar {
-  background: #2f3640;
-  width: 100%;
-  height: 64px;
-}
-
-.App {
-  position: 'relative';
-  width: 100vw;
-  height: 100vh;
-}
-
-.browser-view {
-  position: 'relative';
-}
-
-.page {
-  background: #353b48;
-  height: 100vh;
-  overflow-x: hidden;
-  overflow-y: auto;
-}
-
-.artist-card-name {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.artist-page-image {
-  width: 100px;
-}
-
-.current-song {
-  border-top: thin solid #485460;
-  border-bottom: thin solid #485460;
-}
-
-.lyric-body {
-  white-space: pre-wrap;
-  height: 60vh;
-  overflow-y: auto;
-}
