+++ src/theme/colors.js
@@ -1,0 +1,8 @@
+const colors = {
+    accent: "#FFA600",
+    white: "#FFFFFF",
+    black: "#000000",
+    primary: "#0b1438",
+    primaryDark: "#13172c",
+    background: "#081050"
+}
